+++ src/groups/special_linear.jl
@@ -26,7 +26,7 @@ end
 
 @inline function active_traits(f, ::SpecialLinear, args...)
     return merge_traits(
-        IsGroupManifold(MultiplicationOperation()),
+        IsGroupManifold(MultiplicationOperation(), LeftInvariantRepresentation()),
         IsEmbeddedSubmanifold(),
         HasLeftInvariantMetric(),
         IsDefaultMetric(EuclideanMetric()),
@@ -80,6 +80,9 @@ end
 
 inverse_translate_diff(::SpecialLinear, p, q, X, ::LeftForwardAction) = X
 inverse_translate_diff(::SpecialLinear, p, q, X, ::RightBackwardAction) = p * X / p
+
+adjoint_action!(::SpecialLinear, Y, p, X, ::LeftAction) = copyto!(Y, (p * X) / p)
+adjoint_action!(::SpecialLinear, Y, p, X, ::RightAction) = copyto!(Y, p \ X * p)
 
 function inverse_translate_diff!(G::SpecialLinear, Y, p, q, X, conv::ActionDirectionAndSide)
     return copyto!(Y, inverse_translate_diff(G, p, q, X, conv))
@@ -151,11 +154,4 @@ end
 function Base.show(io::IO, M::SpecialLinear{Tuple{Int},𝔽}) where {𝔽}
     n = get_parameter(M.size)[1]
     return print(io, "SpecialLinear($n, $(𝔽); parameter=:field)")
-end
-
-translate_diff(::SpecialLinear, p, q, X, ::LeftForwardAction) = X
-translate_diff(::SpecialLinear, p, q, X, ::RightBackwardAction) = p \ X * p
-
-function translate_diff!(G::SpecialLinear, Y, p, q, X, conv::ActionDirectionAndSide)
-    return copyto!(Y, translate_diff(G, p, q, X, conv))
 end